refactor(chat): extract assistant workflow timeline view
