+++ CHANGELOG.md
@@ -40,6 +40,7 @@ The format is inspired by [Keep a Changelog](https://keepachangelog.com/en/1.1.0
 - Split history/fork viewer internals into dedicated modules: `history-viewer-view.ts` (overlay rendering/filter UI), `history-viewer-types.ts` (shared row/message contracts), and `history-fork-utils.ts` (fork target mapping + entry-id resolution + naming), while removing the legacy standalone fork-picker overlay path in favor of the unified history viewer flow.
 - Extracted no-project welcome dashboard internals into focused modules: `welcome-dashboard-data.ts` (local inventory discovery + CLI update status aggregation) and `welcome-dashboard-view.ts` (centered welcome rendering), reducing `chat-view` responsibilities and improving decomposition around the no-project experience.
 - Extracted assistant workflow/thinking/tool summarization logic into `src/components/chat-view/workflow-utils.ts` (thinking normalization, standalone-codeblock detection, workflow grouping/collection, tool-call summarization, workflow expansion-state derivation), with `chat-view` now using thin delegating wrappers.
+- Extracted assistant workflow timeline rendering into `src/components/chat-view/assistant-workflow-view.ts`, moving workflow detail-entry rendering and expand/collapse UI templating out of `chat-view` while keeping state orchestration in `chat-view` callbacks.
 - Compaction status rendering was reduced to a minimal workflow-style row with collapsed-by-default details instead of a heavy status card.
 - Auto-rename extension recommendation and desktop config bridge now target `@byteowlz/pi-auto-rename`, with dynamic command-to-package resolution for config-intent slash commands (including `/auto-rename config`) instead of hardcoded package-name routing.
 - Recommended notifications extension now defaults to `pi-smart-voice-notify`, and Packages auto-migrates legacy `pi-desktop-notify` installs by installing the new package and removing the old one.
