+++ src/components/chat-view/workflow-utils.ts
@@ -21,6 +21,7 @@ export interface WorkflowMessage {
 	thinking?: string;
 	errorText?: string;
 	isStreaming?: boolean;
+	isThinkingStreaming?: boolean;
 }
 
 export interface WorkflowToolCallGroup {
